requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
biosppy==1.0.0
grad-cam==1.4.6
matplotlib==3.5.3
numpy==1.23.5
pandas==1.1.5
Pillow==9.3.0
pyhrv==0.4.1
pytorch_lightning==1.8.6
scikit_learn
scipy==1.9.3
tabulate==0.9.0
torch==1.12.1+cu116
torchvision==0.13.1+cu116
torchmetrics==0.11.0
tqdm==4.64.0
wfdb==4.1.0
xgboost==1.7.2
pyentrp==0.7.1
PyWavelets==1.4.1
scikit-optimize